Clean up code and reimplement conversion logic
A lot of the existing code was refactored to provide a cleaner "base" to
build off of. The `Environment` struct now has its own module as it
began to grow in size and will continue to do so as additional trait
implementations are added.

Support for 'staging' and 'production' environments was added as these
are both common and useful names for runtime environments.