Add Axum web framework and improve Config module
The `axum` web framework provides a better developer experience than the
standard library (std) while offering access to powerful middleware via
the `tower` and `tower-http` ecosystem.

The config module was modified to make loading config file sources
simpler and more reliable. File names do not have to be parsed and
validated. Instead, the filename is derived from the current value of
APP_ENV and a compatible file extension is automatically appended to it.

Lastly, the state module provides a straightforward method for
initializing and sharing state, or data, that must be available to
various parts of the application.